cria DTO's para novo SalableBook e altera tipo do campo 'units'

+++ src/main/java/caio/portfolio/livraria/infrastructure/entity/book/salable/SalableBook.java
@@ -31,7 +31,7 @@ public class SalableBook extends Book {
 	private BigDecimal price;
 	
 	@Column(name="units", nullable=false)
-	private BigDecimal units;
+	private Integer units;
 	
 	@Override
 	public boolean equals(Object o) {
